expose IPC in API submodule
