+++ package.json
@@ -19,7 +19,7 @@
   ],
   "pear": {
     "name": "pear-electron",
-    "bootstrap": "pear://0.6311.dqgs5prptqozgwzanes31atj1imzbf5dfg4tgpgkp5mke83oziyy",
+    "bootstrap": "pear://0.6313.dqgs5prptqozgwzanes31atj1imzbf5dfg4tgpgkp5mke83oziyy",
     "stage": {
       "skipWarmup": "true",
       "only": [
